+++ myBackupTrim.ps1
@@ -1,3 +1,5 @@
+#requires -version 5.1
+
 #MyBackupTrim.ps1
 
 #trim full backups to the last X number of files
@@ -14,7 +16,7 @@ Param(
     [string]$Pattern = "*-FULL.rar",
 
     [Parameter(HelpMessage = "Specify the number of the most recent files to keep")]
-    [Validatescript( { $_ -ge 1 })]
+    [Validatescript({ $_ -ge 1 })]
     [int]$Count = 4
 )